Add multi-dataset support, full trajectory fix, and config updates
- Add dataset scripts for EuRoC, FAST-LIVO2, RPNG-AR, UTMM
- Add run scripts for each dataset
- Fix _save_full_trajectory: read from droid_c2w snapshots, clamp tau>=0
- Add trajectory_filler.py (SE3 Lie-algebra interpolation)
- Add configs for EuRoC, FAST-LIVO2, RPNG-AR, UTMM datasets
- Update frontend (dbaf, dbaf_frontend, motion_filter, depth_video)
- Update Gaussian model, utils, loss, vis_utils
- Update loop detection and metric model
- Update middleware_utils and set_env.sh

+++ scripts/frontend_vo/slam/visual_frontends/visual_frontend.py
@@ -1603,7 +1603,7 @@ class RaftVisualFrontend(VisualFrontend):
                        "kf_idx_to_f_idx":     self.kf_idx_to_f_idx,
                        "is_last_frame":       batch["is_last_frame"],
                        # TTD 2024/05/17
-                       "viz_out_idx_to_f_idx": self.cam0_timestamps[self.viz_idx].to(torch.long)
+                       "viz_out_idx_to_f_idx": self.cam0_timestamps[self.viz_idx]  # keep as float for timestamp consistency
                       }
             self.viz_idx[:] = False
         else:
